requirements.txt
python-binance==1.0.15
apprise==0.9.6